requirements.txt
python-telegram-bot==20.3
fastapi
uvicorn
pymongo
pandas
python-dotenv
jinja2

